move definitions from BoardConfig.hpp
required a run of cmake
